Fixed building UAC and UAS dialogs when headers contain parameters
When building a UAC or UAS dialog (pjsip_dlg_create_uac / uas) the values
passed need to be string representations of SIP URIs, not headers, otherwise
the parsing will fail if there are header parameters.

Also, contact header parameters need to be set after building the UAC dialog
so that they are sent out when building the request.