Refactored SDP handling
Each RTP transport will now have the view of the whole SDP plus the
index to which that stream belongs. This allows us to build subsequent
SDPs having PJSIP attach the transport specific information.

This finally allowed us to generate a re-INVITE after ICE negotiation was
done, in order to inform the other endpoint (and any middleboxes) about the
negotiation result.