Fix FrameBufferVideoRenderer so that it can be attached to any producer
Before this change it was possible to change the producer on the fly, but
only to another producer of the same type. A VideoCamera is an active
producer, whereas a RemoteVideoStream is a passive one. Now the renderer
will destroy the underlying video port and recreate it automagically, in
order to accomodate any producer type.