pjsip: fix crash when using certain versions of libav
We are currently using avcodec_get_frame_default for initializing AVFrame
structures. Yes, it's a deprecated structure, but that's another story. We
happen to allocate those frames on the stack. FFMpeg and libav diverge in
what they do internally, as libav might allocate some buffer which needs
to be freed when setting the defaults. The crash was caused because libav
does a pointer comparison, but because we allocate the frame on the stack
but don't zero out the memory, pointers contained garbage.

TLDR: unitialized memory access / freeing because ffmpeg and libav differ
internally, yay!