pjsip: avoid crashes on OSX if video size changes
For some reason, OSX decides to ignore our request for opening the device at
a given resolution and gives us images in a different one. What's even
weirder is that this doesn't happen immediately, it can happen at a later time,
in the middle of a call.

This patch sets the frame size to the correct value, so PJSIP doesn't read
bogus memory if the resolution changes (it always seems to change to a smaller
value).