Fixed issue when sending non-text messages without using CPIM
The code used UTF-8 to decode ANY message it received, which is incorrect
for anything that does not have a text based mime type as well as for text
based mime types that are encoded with something else. In order to solve it
we now attach the charset to the Content-Type header to indicate if the
message needs to be decoded and with what charset.

Alongside the fix, the following changes were made:
 - unified ChatIdentity and CPIMIdentity into ChatIdentity
 - replaced ChatMessage & CPIMMessage with SimplePayload & CPIMPayload
 - the ChatStreamGotMessage notification now provides a Message instance